backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.2.0
mercadopago>=2.0.0
reportlab==4.0.7
# Dependencias para AFIP
cryptography==46.0.3
lxml==6.0.2
requests==2.32.5
# Dependencias para PDFs y QR
qrcode==7.4.2
Pillow==12.0.0
weasyprint==61.2